Fix compile warnings: -Wsign-conversion in Track.h/Song.h/AudioBuffer.h
Stage 3 of plans/warning-cleanup.md, biggest-blast-radius headers
first. Each turned out to be a single source location repeated dozens
of times across every TU that includes the header - e.g.
AudioBuffer.h's 61 raw warnings were all one site, resize()'s memcpy
length.

Two shapes covered every case: getChild(int)/getScene(int)/
getInstrument(int)-style accessors cast at the actual container index
(children_[static_cast<size_t>(i)]) rather than changing the int
parameter, since int is the established convention for these
accessors across the codebase and changing it would ripple into every
caller; a bare arithmetic expression (the memcpy length) just gets a
single static_cast<size_t> on its int operand.

Clean build: 349 -> 125 warnings, all -Wsign-conversion remaining.

Verified: clean rebuild (0 errors), full ctest pass, --render smoke
test.

+++ AudioBuffer.h
@@ -165,7 +165,7 @@ class AudioBuffer final {
     auto new_data = (float *)aligned_alloc(16, getAlignedSize(channels_ * new_size));
     auto frames_to_copy = frames_ < new_size ? frames_ : new_size;
     for (int j = 0; j < channels_; j++) {
-      memcpy(new_data + j * new_size, data_ + j * frames_, frames_to_copy * sizeof(float));
+      memcpy(new_data + j * new_size, data_ + j * frames_, static_cast<size_t>(frames_to_copy) * sizeof(float));
     }
     free(data_);
     data_ = new_data;
